requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
Jinja2==3.1.2
python-dotenv==0.21.1
requests==2.28.2
pyotp
werkzeug
pydantic
httpx
pymongo
kafka-python
Pillow

pydantic[email]
qrcode
slowapi

